refact folders and modules, init pull #4
